[♻️refactor]: 1차 배포를 위한 주요 컴포넌트 리팩토링 및 파일 이동
