--- src/pages/RoomPage/RoomViews/index.ts
@@ -1,3 +1,0 @@
-export { default as MemberList } from './MemberList/MemberList';
-export { default as RoomHeader } from './RoomHeader/RoomHeader';
-export { default as TestInfo } from './TestInfo/TestInfo';
